New Virtual Drip backend based on an emulator for a Yamaha V9958 video display processor along with the support in the Proxy.
Proxy protocol has been extended and enhanced to support the new backend
and its features, including new commands and capabilities specific to
the V9958.

+++ Code/VirtualDrip/src/protocol_debug.h
@@ -15,7 +15,7 @@
 #include <stdint.h>
 
 /** Print payload bytes as uppercase hex, or "none" for an empty payload. */
-void print_payload_hex(const uint8_t *payload, uint8_t length);
+void print_payload_hex(const uint8_t *payload, uint16_t length);
 
 /** Print packet-type-specific details such as data value or key fields. */
 void print_packet_detail(const Packet *packet);
